Updated Schema's for DB and fixed endpoints for Invoices
Nov 11th 8:22PM
Updated the schema and starter inputs to have the new changes made by Will and I.

Added a date check to the checks file in utils.

Fixed the endpoints to use the correct structure with a middle QueryController file.

Fixed endpoints to correctly create the set and get strings for the database.

Thanks Will!