Added/Modified Endpoints and Database
First and foremost, changed the naming convention of the database. I disliked that we couldn't use camelcase, so I changed it so that we could. Only downside is we now need to make sure to use double quotes everytime we try to access the database's stuff.

Then, I edited all the endpoints to reflect this.

Then, I created new files/endpoints for Orders.

I also modified the Invoice endpoints to ONLY return the columns for the invoice, rather than the columns of the entities and orders that were joined into the table for the query.

Changed swagger to reflect all this, and made some small tweaks there!

Changed input data to modify to the new database, and added a few more in for testing purposes.

Lastly, small thing, when calling npm run dev, it now has a string displayed, that takes you straight to the api-docs!