Updated database and endpoints
Added Username column to Entity table, and added PaidStatus, ArrivedStatus, and DeliveredStatus columns to Order table.

Added endpoints for Order to return all orders by the given entity ID, and a second with a given persona as well.

Modified a lot of files and endpoints to work with the database changes.